requirements.txt
beautifulsoup4==4.8.1
discord.py==1.2.3
pytz==2019.3
requests==2.22.0